move replicationcontroller to subpackage of pkg/controller to make room for addition of other controllers
